Removed strange warning

+++ fruithunter/Source/Drawable/Mesh/Mesh.cpp
@@ -334,9 +334,9 @@ float Mesh::castRayOnMesh(float3 rayPos, float3 rayDir) {
 		int length = (int)m_meshVertices.size() / 3;
 		for (int i = 0; i < length; i++) {
 			int index = i * 3;
-			float3 v0 = m_meshVertices[index + 0].position;
-			float3 v1 = m_meshVertices[index + 1].position;
-			float3 v2 = m_meshVertices[index + 2].position;
+			float3 v0 = m_meshVertices[(int)index + (int)0].position;
+			float3 v1 = m_meshVertices[(int)index + (int)1].position;
+			float3 v2 = m_meshVertices[(int)index + (int)2].position;
 			float t = triangleTest(rayDir, rayPos, v0, v1, v2);
 			if ((t > 0 && t < closest) || closest < 0)
 				closest = t;
